Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> (1 comment)
> 
> > Patch Set 7:
> > 
> > > Patch Set 6:
> > > 
> > > (4 comments)
> > > 
> > > > Patch Set 6: Code-Review-1
> > > > 
> > > > (4 comments)
> > > > 
> > > > change goes to the right direction but I'd strongly prefer that you use a common secret template for this
> > > > 
> > > > If you have any troubles let me know and I can help you with that
> > > 
> > > I welcome all help because I don't know how to do that.  Thanks!
> > 
> > Done. Let's see if it works. Let me know if you have any questions.
> > I hope that this simple example will help you to understand the common secret template
> 
> My stack overflowed.  I don't understand where the value of the global secret gets initialized and why it doesn't get changed every time a component deploys (which would be bad).

This is because we use the concept of "masterPassword" (see https://masterpassword.app/masterpassword-algorithm.pdf).

In short. Every deployment with the same master password will have the same passwords generated. This allows us to increase security (because everyone may use different masterPassword for his or her deployment) but in the same time we get reproducibility.

> Also don't know why all the secret(s).yaml files include common.secretFast now

This is the heart of our solution that handles all of the magic. So instead of manually creating secrets you just include this one and fill in "secrets" section in you values.yaml.

Based on this common.secretFast will create kubernetes secrets. Additionally we support concept of "externalSecret" or "bring your own secret" in other words. This means that if "externalSecret" field is evaluated to non-empty string, the component will not create a new secret but instead reference the one that is already existing. And this is actually heavily used here.

So the secret is really created in the top most directory. As "password" field is never set here the actual password is automatically generated by our template using master password algorithm.

Then we pass the name of that secret to all submodules that are using it. In those modules becuase "externalSecret" would evaluate to non-empty string they will not create new secrets but rather reuse the one that is comming from the parent chart

> and whether that causes collisions when some included it previously (I assume for different secrets).

This should be included only once per component, otherwise the deployment will fail

Patch-set: 7
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>