Update patch set 9
Patch Set 9:

> Patch Set 9: Code-Review-1
> 
> Chart onap-dcaegen2
> STATUS: FAILED
> Chart problem
> info:
>   Description: 'Release "onap-dcaegen2" failed: Deployment.apps "onap-dcae-cloudify-manager"
>     is invalid: [spec.template.spec.volumes[7].secret.secretName: Required value,
>     spec.template.spec.containers[0].volumeMounts[7].name: Not found: "cm-secrets"]'
>   first_deployed:
>     nanos: 445971320
>     seconds: 1600123269
>   last_deployed:
>     nanos: 445971320
>     seconds: 1600123269
>   status:
>     code: 4
> name: onap-dcaegen2
> namespace: onap

I did a dry-run deployment of this set of charts and looked at the rendered output.  As the error message above suggests, the secretName for the volume evaluates to the empty string.  (I see this
happening also for other components where the chart is trying to set an environment variable from 
the secret.)   The chart explicitly names all of the secrets, but it is using uid to look up the
secret name.  Could this be the problem?

Patch-set: 9
Reviewer: Gerrit User 511 <511@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>