Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> if you look at here: https://logs.onap.org/onap-integration/gating//112410-2/infrastructure-healthcheck/k8s/onap-helm/onap-certInitializer.log
> we see that this linting is done on purpose ^_^

Nope. This chart is not intended to be deployed as is. Only as a requirement from some other component and I expect that this apply to all our charts that we created to wrap global values

Patch-set: 2
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>