Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 1:
> > 
> > if you look at here: https://logs.onap.org/onap-integration/gating//112410-2/infrastructure-healthcheck/k8s/onap-helm/onap-certInitializer.log
> > we see that this linting is done on purpose ^_^
> 
> Nope. This chart is not intended to be deployed as is. Only as a requirement from some other component and I expect that this apply to all our charts that we created to wrap global values

you're not supposed to take care of your children :P?

in https://gerrit.onap.org/r/c/oom/+/112410, as it's a release name, the name override is not taken into account and so this doesn't apply yes.
But it breaks helm chart deployment (as we see in the logs).

So I prefer to handle the case by lowering everywhere as it'll force our names to always be DNS compatible.

Another solution would be to replace any capital letter in a word by '-lowerOfTheCapitalLetter'

Patch-set: 2
Reviewer: Gerrit User 2319 <2319@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>