Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> In general I agree to merge both entry point scripts, this would need some time for testing and intensive review. @Dan Should we give it a try?

I would prefer to merge the entry point scripts in a separate change. This seems fairly risky, so
we need to take the time needed to test it - as you say.

 My understanding is that this fix must be done before we can merge any other SDNC changes, including our version bump to use the Guilin version of our dockers (we're currently still using Frankfurt)

Sylvain - are you okay with merging this as is, and then handling combination in a separate change.

Patch-set: 1
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>