Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> > Patch Set 11:
> > 
> > Sylvain - Can you help getting the error/log for Cloudify pod deployment?
> 
> @Vijay: please tell me where to find the relevant logs and I'll try to catch them

I notice that policy, sdc, and dmaap are all failing at the Helm deploy stage.  Because of that, DCAE will not come up fully (policy handler and service change handler will be stuck at the init stage).

In addition, the Cloudify Manager pod is starting, but not becoming ready.  The stdout/stderr log is interesting, because it does NOT show the final step of actually starting the whole Cloudify system.  The log shows the script copying the database information from the container file system onto a persistent volume but nothing after that.   There should be some file linking after that, and then control is turned over to Cloudify with an /sbin/init command.  It's possible that the log isn't complete, but it's also possible that something is happening with the copy operation.

One thing to check is whether there is a file called "load-plugins.log" in the root directory of the CM container. (Does "/load-plugins.log" exist?).  If so, what does it contain?

I'm using the Wind River environment (but not the same k8s cluster as Vijay), and I saw a successful deployment of patchset 10--with a limited deployment that did not include policy and sdc. (And also not the DCAE policy handler and service change handler.)

Patch-set: 11
Reviewer: Gerrit User 511 <511@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>