Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> it seems that the new image from SDNC needs more RAM (https://logs.onap.org/onap-integration/gating//112626-1/infrastructure-healthcheck/k8s/kubernetes-status/pod-onap-sdnc-0-sdnc-logs.html)
> please also update request/limits

Looking into this ... I think actually the problem has to do with the file used to tune Java settings (setenv), so the fix will almost certainly include a new version of that file.  Also will verify that request/limits match up with our tuning (we need to tune heap size on the java command line, and also make sure we're not asking for more heap than the limit).

Patch-set: 1