Update patch set 1
Patch Set 1:

@Morgan,

OSDF healthcheck is sending "content-type" for a get request which is throwing 400. OSDF healthceck was implemented from the SNIRO simulator(A placeholder for OOF in one of the early releases). I guess it is time that we rename the healthcheck name to OOF and request format. Is it okay to do so it in the testsuite?

Please let me know your views.

Patch-set: 1
Reviewer: Gerrit User 2671 <2671@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>