Update patch set 6
Patch Set 6:

FYI - I have another change to bump the CDS versions (https://gerrit.onap.org/r/c/oom/+/113061), which is dependent on this one.   The chained CI is running on that now.   Wasn't sure if anyone was planning to rerun the CI build on this one to take a look at the SDC test issue, but wanted to mention that other review in case you wanted to wait for that to finish to see if the issue occurs there too. It shouldnt have anything to do with either change (neither of them touch the SDC charts or their dependencies)

Patch-set: 6
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>