Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Krishna, this patch should add STDOUT logs to all OOF HAS components or only api one?

The changes in the log.conf will be applied to all the OOF HAS components. But we are running HAS-api in a different way, it is again redirecting into a file. So, I had to make that change in the deployment for HAS-api. Overall, the patch is for all OOF-HAS components. The log.yml is for OOF-OSDF. THe logback files in the CMSO is already configured for STDOUT. There are no changes necessary for it.

Patch-set: 1