Update patch set 1
Patch Set 1:

While I understand that this change was a leap forward in right direction I'd like you to consider changing the implementation as the current one, as I underlined in commit msg, breaks one of the architectural principles and fails onap deployment. It went unnoticed in any of integration testing as they are performed online which is not entirely "target" deployment model for ONAP. The bug was discovered in Samsung lab where we do perform deployment in offline manner.

Patch-set: 1