Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Instead of this change I propose to use any ONAP image to perform keytool and openssl operations.

Yes, sure, I'm not insisting on reverting if a better solution will be provided soon, just wanted to drag everyone's attention to the issue.

Just please beware that your proposal has the side effect that should that image, that would get/already is defined in one of the OOM charts, be a subject to change (removal, tag change) the cert mgmt mechanism will again fail in offline deployment. So there would have to be some kind of sync mechanism between what's defined in OOM charts in scope of images and here in the docker cmd spawning the container based on that image.

Patch-set: 1