Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > Instead of this change I propose to use any ONAP image to perform keytool and openssl operations.
> > 
> > Yes, sure, I'm not insisting on reverting if a better solution will be provided soon, just wanted to drag everyone's attention to the issue.
> > 
> > Just please beware that your proposal has the side effect that should that image, that would get/already is defined in one of the OOM charts, be a subject to change (removal, tag change) the cert mgmt mechanism will again fail in offline deployment. So there would have to be some kind of sync mechanism between what's defined in OOM charts in scope of images and here in the docker cmd spawning the container based on that image.
> 
> Hi Bartek, I'm aware about this. I don't know exactly how offline installer works but I suspect that a crawler extracts list of ONAP images from some place. Do you know from what place exactly?

You are correct. We have a script for that which is a part of toolchain for building a deployment package: https://git.onap.org/oom/offline-installer/tree/build/creating_data/docker-images-collector.sh

What it does is compiling charts (make) and then simply parses the output of resolved helm templates, see line #63.

But please do not consider this whole issue from Samsung's deployment model, I just mentioned it for reference. The real issue with this change is that it's against some architectural concepts (which offline-installer merely follows) which I believe we ought to follow.

Patch-set: 1