Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> all logs from pods are almost empty
> @krishna, is it normal?
> 
> by the way, what's "mso-simulator" pod? (https://logs.onap.org/onap-integration/gating//113414-1/infrastructure-healthcheck/k8s/kubernetes-status/pod-onap-oof-cmso-service-7cd4c5c949-kjlqb-mso-simulator-logs.html)
> Are you sure we need to deploy it?

I have to check with the developers who contributed the code to CMSO. They are currently not active in the community.

I beleive the simulator is mocking the SO and it has some robot test suites of CMSO. Since no use case is currently supporting it, the integration with SO is not done and there is no contributions to support it.

We can remove the simulator and try pointing it to the real SO, assuming that it will not break any integration test suites. It should work if there are no tests running for CMSO.

Patch-set: 1
Reviewer: Gerrit User 2671 <2671@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>