Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> (1 comment)
> 
> You created the config map two times (with same name, just the app label changed) and mounted it 2 times at the same place into the pod.
> I believe one is sufficient no?
> 
> I'm OK with having them on OOM, I'd just like to have @Mike, @Borislav and @Krzysztof thoughts

I'm sorry for the confusion. The difference in the names is nearly invisible, and I didn't write up any description for needing the second ConfigMap.  I don't blame you for calling me out.  😊

During the work, I discovered a second initialization file: identities.dat, and I did not alter the jira ticket to discuss.  This file is not for the database, but had been built into the aaf-cass image for some time.  I suppose I can look into mounting it on other containers would access it.

This second file and the set of database initialization files are in different ConfigMaps, with different names, and mount points.  It is difficult to see the difference in the names (one ends in 's', the other in 'a').  These unfortunate names are used within the image for different directories.  


John

Patch-set: 1