Update patch set 1
Patch Set 1:

And, I didn't pay close enough attention to the label.  I cut-n-paste from another project and I do not know of the policy for why the labels should be one way or the other.

However, I'll change the app label to be the same for both config maps.

Thanks.

Patch-set: 1