Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> (2 comments)
> 
> I didn't looked at all the patch as I'm not sure to understand exactly why we need it and why we cannot use some ONAP APIs in order to achieve that.
> 
> Furthermore, it mandates Internet access in order to use it if I understand well and thus it may break offline installation.
> 
> So, README is interesting (and I really appreciate that you gave right documentation) but I'd like you to show me where you desperately needs this and why you cannot do it another way.

Sylvain,

Well, the whole idea here is to be able to generate the configMap, which contains 3GPP openAPI descriptions. These are changing ~3-4 times a year, and we`d like to have a possibility to load such openAPI descriptions to the relevant DCAE components, which might use them.

There are some openAPIs delivered already within the packages, but we`d like to make it easier for operators to prepare a configMap with updated openAPIs using a script (and not manually downloading and preparing a public/local URL map), then load the ready-made configMap with another one.

The target solution is openAPi onboarding in SDC, but this is something we could do only in Honolulu release. We agreed the solution with Vijay, a question is only, where to place these scripts, so they`re available to the operator for use.

Rgds,
Damian

Patch-set: 2
Reviewer: Gerrit User 5597 <5597@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>