Update patch set 2
Patch Set 2:

Thank you Damian for Your thorough explanation, I really appreciate it.

However, in my understanding (anyone please prove me wrong if that's the case) ONAP is a platform that is meant to be _independent_ of the underlying OS layer - Kubernetes being it's only and canonical driver. After reading the readme and the sentence "Script works and has been tested on Ubuntu 18.04 and Ubuntu 19.10 which covers actual ONAP OS environment" (I use RedHat/Centos) and Sylvain comment to it I not only support his comment but then I even think that this tool doesn't even belong to OOM project because it doesn't leverage on the underlying Kubernetes platform (runtime wise).

The external repository is in that case not the only dependency here then but the OS on which these scripts can run is also a constraint hence a dependency as such.. Even if the only script to be run on target env on which ONAP is deployed doesn't need any artifacts from Internet it's still constrained by OS it's compatible with. And it's not possible to test each corner cases here, in OS regard.

Hence my suggestions:
1.Treat these tool as auxiliary and put it in some other repository (one of integration repos? not sure).
2.Deploy it as a native kubernetes application, i.e. put all required artifacts to a relevant image (new one?) and allow running the tools on Kubernetes platform (as pod) on demand. In that case: no external dependencies needed, all artifacts contained in a local image and no constrain from scripts perspective on underlying OS. In real life I think that ONAP operator would like to run any such tool on target deployment in this scenario.

Patch-set: 2
Reviewer: Gerrit User 5097 <5097@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>