Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> I believe we have an issue on the onboarding:
> 
> #### Certificate Authorization Artifact
> 2020-10-07T23:27:53.846+0000 INIT [cadi] cadi_keyfile points to /opt/app/osaaf/local/org.onap.nbi.keyfile
> 2020-10-07T23:27:53.865+0000 INIT [cadi] https.protocols set by cadi_protocols in CADI Properties
> 2020-10-07T23:27:53.865+0000 INIT [cadi] jdk.tls.client.protocols set from Default Protocols
> 2020-10-07T23:27:54.893+0000: Trans Info
>          Read Artifact 454.5089ms
> org.onap.aaf.cadi.CadiException: org.onap.aaf.cadi.LocatorException: No Entries found for 'https://aaf-locate.onap:8095/locate/onap.org.osaaf.aaf.cm:2.1'
> 	at org.onap.aaf.cadi.aaf.v2_0.AAFConHttp.rclient(AAFConHttp.java:135)
> 	at org.onap.aaf.cadi.aaf.v2_0.AAFCon.client(AAFCon.java:227)
> 	at org.onap.aaf.cadi.configure.Agent.readArtifact(Agent.java:558)
> 	at org.onap.aaf.cadi.configure.Agent.main(Agent.java:321)
> Caused by: org.onap.aaf.cadi.LocatorException: No Entries found for 'https://aaf-locate.onap:8095/locate/onap.org.osaaf.aaf.cm:2.1'
> 	at org.onap.aaf.cadi.aaf.v2_0.AbsAAFLocator.best(AbsAAFLocator.java:239)
> 	at org.onap.aaf.cadi.aaf.v2_0.AAFConHttp.rclient(AAFConHttp.java:133)
> 	... 3 more
>  
> #### Certificate Authorization Artifact must be valid to continue

The above implies that the services are not starting.  In testing this change, I looked at the cassandra container only.  The data is the same and I didn't try to startup any of the services before I made this change request.  I didn't anticipate a problem as follows.  This can be related to what you are seeing


When I run the helm templates for aaf services, the containers' image are not correct.   This issue exist on the master branch before my change.  I use 'helm template --debug  aaf/components/aaf-locate | less'

From the output, in Deployment for the locate service (as example), I see this:

      containers:
      - name: aaf-locate
        workingDir: /opt/app/aaf
        command: ["bin/locate"]
        image: nexus3.onap.org:10001/
        imagePullPolicy: Always

The string for image is wrong.  It should be similar to the one like for aaf-config container:
     - name: aaf-locate-config-container
        image: nexus3.onap.org:10001/onap/aaf/aaf_config:2.1.23


This template issue is recreated in the parent commit of my change.   I think perhaps something is different with the common templates which require a value  which aaf isn't setting now?

What is fix?

Patch-set: 2