Update patch set 14
Patch Set 14:

> Patch Set 13:
> 
> Dear Seshu, 
> As an author of https://gerrit.onap.org/r/c/oom/+/113072 I have question related to this ticket:
> 
> - I saw, that every component contains entries required to use cert-initializer; Most of them use it. SO-Monitoring also contains such entries (with slight differences) but cert-initializer is discarded to be used. Can you explain why it is done? 
> SO-Monitoring was prepared to use https here -> https://gerrit.onap.org/r/c/so/+/111540 [SO project] 
> This implementation of cert-initializer may strongly influence on my ticket.

Hi, once this patch was made, so-monitoring was http only and thus was not needing certificates

For https://gerrit.onap.org/r/c/oom/+/113072, I believe you should rebase it on top of this one.

For the merge part, I'd like to have :
* basic vm to work
* 1 (preferably 2) OOM commiters review, excluding me as I'm one of the author.

I know this has stuck SO for a while but we need to be very careful with SO (and AAI, SDC, SDNC) and I believe I was not enough careful the last times.

Patch-set: 14
Reviewer: Gerrit User 2319 <2319@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>