Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Hi Bartek,
> > Well, the sentence is related to the fact, that the script was tested on these Ubuntu versions.
> > We haven`t introduced any OS specifics in a script, and due to this, we decided to code it in bash, even though Python/Java will be easier, but I`d bring additional compatibility issues.
> > 
> 
> by the way, Bash is GPLv3 so we're not allowed to use it. 
> You can move to sh / dash / zsh (but then you'll need to check OS compatibility for the two last I guess) if you want but I think Bartek proposal (having a Job doing that) would be for the best no?
> 
> And I'm OK to ask for a new repo in order to embed scripts / docker generation + add something in "main OOM" in order to launch it.

Ok, we`d check with sh only.
We cannot use this as a job, as the intention is, that the operator prepares a configMap as he/she needs. If we`d put that script in the job, then we`d create an internet dependency, which we want to avoid.

Let`s discuss further in the OOM call on Wednesday.

Patch-set: 2
Reviewer: Gerrit User 5597 <5597@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>