[SDNC] Fix typo in setenv script
The setenv script used to set garbage collection options in Frankfurt
contains a typo.  Unfortunately, since the GC logging options are different
in Frankfurt vs Guilin, we cannot simply cherry pick the Guilin change that
fixed this - we must submit a separate change directly to Frankfurt.

This change changes the setenv script so that it will be identical to that
used in Guilin - the differences are now in sdnc values.yaml. There is
a new setting (gcLogOptions), which is set here to the proper values
to enable GC logging for Frankfurt.

Issue-ID: SDNC-1377
Signed-off-by: Dan Timoney <dtimoney@att.com>
Change-Id: Ibacdb598cc561d64d1e197d45fe485a4701e3dd8