[CLAMP] Give times to mariadb to come up
On some environments, clamp database is restarted in the middle of the
init script.
The consequence is then a weird behavior, especially if the users are
not set.
This patch adds more time to the DB in the environment override file
(which is supposed to be used in slow environnement).

Issue-ID: OOM-2597
Signed-off-by: Sylvain Desbureaux <sylvain.desbureaux@orange.com>
Change-Id: Ieb700afbb27610dbad18d860b7fb85ac95c2bb0d