Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > I would personnally prefer to add the needed container image name "somehow" to the mandated list of image that must be downloaded if offline
> 
> > and then we could use this to both fake master password instead of setting it explicitly in the offline installer but also use it here in this fake deployment
> 
> How would you like to include those in ONAP architectural requirements? Because that document and not Offline-installer describes the concepts for the platform. Besides - allowing one exception from specification is a dangerous precedent that anyone in the future will be adducing to justify other "WOs".

Hmmm then where is the offline installer current architecture and processing documented?

BTW
Having a flag which informs the source code which "compiler" is processing the source code is a common practice.

Take a look at gcc and g++ and a typical ifndef in the source code to use different constructions depending if we are building with C or C++ compiler

Here we would use the same concept and just define a flag if the source code is being process by the offline istaller

Patch-set: 2
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>