Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > I would personnally prefer to add the needed container image name "somehow" to the mandated list of image that must be downloaded if offline
> > 
> > > and then we could use this to both fake master password instead of setting it explicitly in the offline installer but also use it here in this fake deployment
> > 
> > How would you like to include those in ONAP architectural requirements? Because that document and not Offline-installer describes the concepts for the platform. Besides - allowing one exception from specification is a dangerous precedent that anyone in the future will be adducing to justify other "WOs".
> 
> Hmmm then where is the offline installer current architecture and processing documented?

It don't see how Your comment relates to my statements but please consider the fact that offline-installer implementation is in no way __reference__ implementation for ONAP offline deployment which seems to be silent assumption in this discussion.

Instead we should completely forget it and focus on avoiding introducing any kind of runtime dependencies. If this requirement is no longer valid please escalate it to PTL to remove point 4.4.7 from arch principles.

4.4.7 says "ONAP must be runnable in a completely private environment with no routing to Internet. This effectively means that any runtime downloading of libraries, definitions and resources from the internet is prohibited and everything must be packaged to the images of the ONAP installation package."

In particular it doesn't say "...everything must be packaged to the images of the ONAP installation package _unless_ there is some fancy workaround in the code for online parts which offline deployment modules can hook into".

Patch-set: 2