[SO] Import various CAs in truststore
Per default, SO truststore has only one CA, the ONAP one.
But we also need MSB root CA. The process to onboard was broken and this
patch solves it
We also needs "common root CAs" in order to discuss with other
components such as the underneath OpenStack.
In this patch we also import all "known" root CA from truststoreONAPall.

Issue-ID: OOM-2606
Issue-ID: OOM-2607
Signed-off-by: Sylvain Desbureaux <sylvain.desbureaux@orange.com>
Change-Id: Ia67bd4aec7a0b122fb9fda11e1e48c4e6e55430c