Update patch set 10
Patch Set 10:

I would rather stay with patchset 9.
As recent observation showed, "if's" around keystore made, that some of preloaded certs were loaded, 
so effectively "aafEnabled" in "if" gave "false" result.
So this changes may result the same, in old cert loading or exposing http port for the environment? 
Not sure if we want such effect.

Patch-set: 10