Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> > Patch Set 10:
> > 
> > I would rather stay with patchset 9.
> > As recent observation showed, "if's" around keystore made, that some of preloaded certs were loaded, 
> > so effectively "aafEnabled" in "if" gave "false" result.
> > So this changes may result the same, in old cert loading or exposing http port for the environment? 
> > Not sure if we want such effect.
> 
> but what happens then when you don't have AAF?

suppose, pod will fail. Up to now SO-Monitoring was hidden, because it has no https exposed ports. 
There is a question what is better - failed container with functionality hidden, like up to now, 
or http exposed port and possible vulnerabilities? 
Another question is, why "aafEnabled" seem to be "false" even if we see aaf is working?

Patch-set: 10