Update patch set 10
Patch Set 10: Code-Review+1

Dears,

I think, we`re good to merge this one.
We see pnfRegistration running properly. basic_Vvm showing issues in So-Openstack-Adapter - which might be related to the infra. SO-Monitoring will help as a troubleshooting tool to see, what has broken in the SO flows ( it exactly shows SO BPMN execution progress).

We can as well merge SO 1.7.5 first, and the rebase and merge this one).

Patch-set: 10
Reviewer: Gerrit User 5597 <5597@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1