Update patch set 10
Patch Set 10:

Ok, as we have seen HTTP/500 received from SO API twice in a row, let me try to revert all containers of SO to 1.7.6, and leave only the SO-BPMN-INFRA as 1.7.7.
This would be exactly same set-up, which we had yesterday, when a successful test was executed.

Patch-set: 10
Reviewer: Gerrit User 5597 <5597@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>