Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-2
> 
> This cannot be changed in this way.
> 
> Init container has to stay where it is.
> 
> If you need to modify this file you need to add additional container what will update the file on the volume when CM has changed but the init container has to stay where it is.
> 
> BTW
> You probably also need to notify the application that its config file has changed no?

Well, application is as we've already checked reacting on its own on config file change. I can keep original init container together with current sidecar, but it appeared to me as unnecessary complication if sidecar container is enough to work.

I'm fine to keep original init container coupled with the one I've added (it's as you see pretty same as init container but with continuous polling set up) if you believe it's necessary.

Patch-set: 3